flask-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
Werkzeug==3.0.1

# OCR
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.1.0

# NLP & BERT
spacy==3.7.2
scispacy==0.5.4
transformers==4.36.0
torch==2.1.0

# Database
pymongo==4.6.1

# Authentication & Security
bcrypt==4.1.2
python-dotenv==1.0.0

# Payments & Notifications
stripe==7.6.0
smtplib

# Email
flask-mail==0.9.1

# Date/Time
python-dateutil==2.8.2

# Validation
marshmallow==3.20.1
